+++ src/main/java/tech/deplant/java4ever/binding/generator/TypeReference.java
@@ -1,15 +1,12 @@
 package tech.deplant.java4ever.binding.generator;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import tech.deplant.commons.Strings;
 import tech.deplant.java4ever.binding.EverSdkContext;
 import tech.deplant.java4ever.binding.SubscribeEvent;
-import tech.deplant.java4ever.binding.generator.javapoet.ArrayTypeName;
-import tech.deplant.java4ever.binding.generator.javapoet.ClassName;
-import tech.deplant.java4ever.binding.generator.javapoet.ParameterizedTypeName;
-import tech.deplant.java4ever.binding.generator.javapoet.TypeName;
+import tech.deplant.javapoet.*;
 import tech.deplant.java4ever.binding.generator.jtype.SdkObject;
 import tech.deplant.java4ever.binding.generator.reference.*;
-import tech.deplant.java4ever.utils.Strings;
 
 import java.math.BigInteger;
 import java.util.List;