+++ src/main/java/tech/deplant/java4ever/binding/generator/jtype/SdkObject.java
@@ -1,6 +1,6 @@
 package tech.deplant.java4ever.binding.generator.jtype;
 
-import tech.deplant.java4ever.binding.generator.javapoet.TypeSpec;
+import tech.deplant.javapoet.TypeSpec;
 import tech.deplant.java4ever.binding.generator.reference.ApiType;
 
 public interface SdkObject {