+++ src/test/java/tech/deplant/java4ever/unit/ConfigTests.java
@@ -10,12 +10,25 @@ import org.junit.jupiter.api.parallel.ExecutionMode;
 import tech.deplant.java4ever.binding.Client;
 import tech.deplant.java4ever.binding.EverSdkContext;
 import tech.deplant.java4ever.binding.EverSdkException;
+import tech.deplant.java4ever.binding.loader.DefaultLoader;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @Execution(ExecutionMode.CONCURRENT)
 public class ConfigTests {
+
+	@Test
+	public void check_ever_sdk_version() throws JsonProcessingException, EverSdkException {
+		assertEquals(DefaultLoader.EVER_SDK_VERSION, Client.version(EverSdkContext.builder().buildNew()).version());
+	}
+
+	@Test
+	public void context_id_increments_as_we_are_using_the_same_lib_for_each_context() throws JsonProcessingException, EverSdkException {
+		assertEquals(1, EverSdkContext.builder().buildNew().id());
+		assertEquals(2, EverSdkContext.builder().buildNew().id());
+		assertEquals(3, EverSdkContext.builder().buildNew().id());
+	}
 
 	@Test
 	public void network_config_from_json_equals_result_config() throws JsonProcessingException, EverSdkException {
@@ -36,8 +49,8 @@ public class ConfigTests {
 				.setConfigJson(configJson)
 				.buildNew();
 		Client.version(ctx);
-		assertEquals("2.3.0", Client.config(ctx).binding().version());
-		assertEquals("java4ever", Client.config(ctx).binding().library());
+		assertEquals(DefaultLoader.BINDING_LIBRARY_VERSION, Client.config(ctx).binding().version());
+		assertEquals(DefaultLoader.BINDING_LIBRARY_NAME, Client.config(ctx).binding().library());
 	}